refactor: extract shared ReportRenderContext and report-data types
run/src/lib/report.ts's BuildReportMarkdownOptions and WriteReportOptions
each redeclared the same four run-step-context fields; extract a
ReportRenderContext base interface they both extend.

Also add core/lib/report/report-data.ts as the canonical ReportData/
ReportSections/DeniedEntry shape for report.js's JSON output, shared by
run/src/lib/report.ts (Report is now an alias of it) and
report/src/lib/command-log.ts (DeniedEntry), instead of each side
independently declaring its own near-identical copy. report/src/main.ts's
previously-untyped `report` variable is now typed as ReportData.
